Show view dialog after creating a new payment
