+++ budgetmanagerpwa/src/components/PaymentForm.jsx
@@ -19,6 +19,10 @@ import {
 } from '../redux/apiSlice'
 import dayjs from 'dayjs'
 import 'dayjs/locale/en-gb'
+import customParseFormat from 'dayjs/plugin/customParseFormat'
+import { getPaymentTitle } from '../redux/utils'
+
+dayjs.extend(customParseFormat)
 
 export default function PaymentForm({
   payment,
@@ -43,7 +47,8 @@ export default function PaymentForm({
   const onFormSubmit = (formData) => {
     if (payee == null) alert('Missing payee')
     else {
-      onSubmit({ payee: payee.id, ...formData })
+      formData.date = dayjs(formData.date, 'DD/MM/YYYY').format('YYYY-MM-DD')
+      onSubmit(payment, { payee: payee.id, ...formData })
       onClose()
     }
   }
@@ -52,7 +57,7 @@ export default function PaymentForm({
       open={open}
       onClose={onClose}
       onSubmit={onFormSubmit}
-      title={title}
+      title={title ? title : payee ? getPaymentTitle(payment, payee) : 'title'}
     >
       <List>
         <ListItem>
