+++ budgetmanagerpwa/src/redux/utils.ts
@@ -8,3 +8,11 @@ export const showUserDetails = (user: User | null) =>
         : user.first_name
       : user.username
     : 'Administrator'
+
+export const getPaymentTitle = (
+  payment: { amount: number },
+  payee: { name: string }
+) =>
+  `${Math.abs(payment.amount)} ${payment.amount > 0 ? 'from' : 'to'} ${
+    payee.name
+  }`
